Rename "atom-items" to "items"
