+++ lib/atom/pane-item.js
@@ -5,7 +5,7 @@ import {CompositeDisposable} from 'event-kit';
 
 import URIPattern, {nonURIMatch} from './uri-pattern';
 import RefHolder from '../models/ref-holder';
-import StubItem from '../atom-items/stub-item';
+import StubItem from '../items/stub-item';
 import {createItem, autobind} from '../helpers';
 
 /**
