+++ lib/controllers/root-controller.js
@@ -17,7 +17,7 @@ import Commands, {Command} from '../atom/commands';
 import GitTimingsView from '../views/git-timings-view';
 import GithubTabController from './github-tab-controller';
 import FilePatchController from './file-patch-controller';
-import IssueishPaneItem from '../atom-items/issueish-pane-item';
+import IssueishPaneItem from '../items/issueish-pane-item';
 import GitTabController from './git-tab-controller';
 import StatusBarTileController from './status-bar-tile-controller';
 import RepositoryConflictController from './repository-conflict-controller';
