+++ lib/github-package.js
@@ -13,7 +13,7 @@ import WorkdirContextPool from './models/workdir-context-pool';
 import Repository from './models/repository';
 import StyleCalculator from './models/style-calculator';
 import RootController from './controllers/root-controller';
-import StubItem from './atom-items/stub-item';
+import StubItem from './items/stub-item';
 import Switchboard from './switchboard';
 import yardstick from './yardstick';
 import GitTimingsView from './views/git-timings-view';
