+++ lib/items/issueish-pane-item.js
