+++ lib/items/issueish-tooltip-item.js
