+++ lib/items/stub-item.js
