+++ lib/items/user-mention-tooltip-item.js
