+++ lib/views/github-dotcom-markdown.js
@@ -5,8 +5,8 @@ import ReactDom from 'react-dom';
 import PropTypes from 'prop-types';
 
 import {handleClickEvent, openIssueishLinkInNewTab, openLinkInBrowser, getDataFromGithubUrl} from './issueish-link';
-import UserMentionTooltipItem from '../atom-items/user-mention-tooltip-item';
-import IssueishTooltipItem from '../atom-items/issueish-tooltip-item';
+import UserMentionTooltipItem from '../items/user-mention-tooltip-item';
+import IssueishTooltipItem from '../items/issueish-tooltip-item';
 import {autobind} from '../helpers';
 
 export default class GithubDotcomMarkdown extends React.Component {
