+++ lib/views/issueish-link.js
@@ -4,7 +4,7 @@ import {shell} from 'electron';
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import IssueishPaneItem from '../atom-items/issueish-pane-item';
+import IssueishPaneItem from '../items/issueish-pane-item';
 
 // eslint-disable-next-line no-shadow
 export default function IssueishLink({url, children, ...others}) {
